Set Session Parameters
IBM's AppScan security software flagged the session because a base path was not set to restrict the session to the login script, and because HTTPOnly was not set.

This commit adds a new config file where custom session options can be added, and loads those settings as part of the Zend bootstrapping.

The install.md document was modified to reflect the change that needs to be made in the session.ini file