Added 403.php
Added a 403 page so that system administrators have the option of setting 403 (forbidden) to a 404 or leaving it the default 403.

By setting the 403 forbidden error to a 404 not found error we are effectively hiding any directories that someone does not have access to view. This will prevent automated security auditing software, such as IBM's AppScan, to flag a security warning for hidden directories.